Make slide numbers fade in
Slide numbers now fade in instead of appearing out of the blue. The
delay is set in a way that the slide number doesn't start appearing
before the counter has already changed to 2 after the title slide. When
going back to the title slide (or when disappearing somewhere else in
the presentation), the slide numbers fade away immediately.