Touch controls!
Using `:touch-controls: true` in the RST header, you can include
controls for touch devices. There are buttons for the next and the
previous page at the left and right screen edges (which can be hidden by
CSS) and left and right swipes are bound to the next and previous
functions as well.
Touch controls disable the regular click actions (i.e. clicking a slide
will go to that slide), because they interfered with the swipe gestures.